third_party/blink/web_tests/http/tests/security/cors-rfc1918/internal-to-internal-fetch.html

<!doctype html>
<html>
<head>
    <!-- No 'treat-as-public-address' CSP directive, so we count as internal. -->
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
    <script src="./resources/preflight.js"></script>
</head>
<body>
    <script>
        test(t => { assert_equals(document.addressSpace, "local"); }, "Sanity check: locality.");

        promise_test(function (t) {
            return fetch(preflightURL('fail-with-500', 'json'))
                .then(function (response) {
                    return response.json();
                }).then(function (json) {
                    assert_equals('success', json.jsonpResult);
                });
        }, "'fetch()' should not generate a preflight for simple requests from internal to internal");
    </script>
</body>
</html>
